refactor cache functions
